Allow foreground and background to be nil.
This allows faces to return nil when looking at either their foreground
or background colors.  mode-icons will look for the specified face,
the mode-line face, the default face and if none of then have a
background or foreground assume either black (foreground) or
white (background).

This should fix the issue reported here:

https://github.com/mattfidler/tabbar-ruler.el/issues/38