Use a more accurate way to determine height for icons
Instead of trying to use the default text height of the current window (or some
other window), ask for the text height of the mode-line face. This should fix
the height when a theme changes the height of the mode-line text to a different
height than the default text height.